requirements/development.txt
-r base.txt

django-extras
django-extensions
coverage
Werkzeug==0.10.1
factory-boy
